openai-protocol: split cache_control normalization into its own module
requests.py crossed the repository line budget; the OpenCode cache-hint
validation and stripping move verbatim into cache_control.py as one
cohesive seam, with the swept echoed-status alias restored where the
decoders use it. No behavior change.
