+++ exp/runtime/openai_protocol/requests.py
@@ -15,7 +15,6 @@ from pydantic import (
     JsonValue,
     TypeAdapter,
     ValidationError,
-    field_validator,
     model_validator,
 )
 
@@ -31,6 +30,10 @@ from exp.runtime.gateway.contracts import (
     GatewayRequest,
     GatewayToolDefinition,
     StructuredTextFormat,
+)
+from exp.runtime.openai_protocol.cache_control import (
+    EphemeralCacheControl,
+    drop_opencode_cache_control,
 )
 from exp.runtime.openai_protocol.errors import OpenAIProtocolError, invalid_field, unsupported_field
 from exp.runtime.openai_protocol.manifest import (
@@ -55,25 +58,6 @@ class _WireModel(BaseModel):
     """Strict private OpenAI wire model used only after official shape validation."""
 
     model_config = ConfigDict(extra="forbid")
-
-
-class _EphemeralCacheControl(_WireModel):
-    """OpenCode/Anthropic cache breakpoint accepted only so it can be dropped.
-
-    The object form is ``{"type": "ephemeral"}`` with an optional ``ttl`` of
-    ``5m`` or ``1h``. An explicit ``ttl: null`` is not in that allowlist.
-    """
-
-    type: Literal["ephemeral"]
-    ttl: Literal["5m", "1h"] | None = None
-
-    @field_validator("ttl", mode="before")
-    @classmethod
-    def _reject_null_ttl(cls, value: object) -> object:
-        """Reject an explicit null TTL while still allowing the key to be omitted."""
-        if value is None:
-            raise ValueError("ttl must be 5m or 1h when present")
-        return value
 
 
 _EchoedItemStatus = Literal["in_progress", "completed", "incomplete"]
@@ -113,7 +97,7 @@ class _AssistantToolCall(_WireModel):
     id: str = Field(min_length=1, max_length=256)
     type: Literal["function"] = "function"
     function: _FunctionCall
-    cache_control: _EphemeralCacheControl | None = None
+    cache_control: EphemeralCacheControl | None = None
 
 
 class _Message(_WireModel):
@@ -422,7 +406,7 @@ def decode_chat(
     Raises:
         OpenAIProtocolError: The body is invalid, unknown, or unsupported.
     """
-    payload = _drop_opencode_cache_control(payload)
+    payload = drop_opencode_cache_control(payload)
     _validate_manifest(payload, CHAT_MANIFEST)
     # The installed SDK's effort literal lags the newest provider tier
     # ("ultra"), so the strict wire model owns reasoning validation.
@@ -557,123 +541,6 @@ def decode_responses(
     except ValidationError as exc:
         raise _validation_protocol_error(exc) from exc
     return DecodedGatewayRequest(alias=request.model, request=canonical)
-
-
-def _drop_opencode_cache_control(payload: JsonObject) -> JsonObject:
-    """Remove supported OpenCode ``cache_control`` annotations from Chat messages.
-
-    Args:
-        payload: Parsed Chat Completions body.
-
-    Returns:
-        The original payload, or a shallow copy whose messages no longer carry
-        a supported ``cache_control`` annotation.
-
-    Raises:
-        OpenAIProtocolError: A ``cache_control`` value is malformed or unsupported.
-    """
-    raw_messages = payload.get("messages")
-    if not isinstance(raw_messages, list):
-        return payload
-    cleaned_messages: list[JsonValue] = []
-    changed = False
-    for index, raw_message in enumerate(raw_messages):
-        message, message_changed = _without_message_cache_control(raw_message, index)
-        cleaned_messages.append(message)
-        changed = changed or message_changed
-    if not changed:
-        return payload
-    cleaned_payload = dict(payload)
-    cleaned_payload["messages"] = cleaned_messages
-    return cleaned_payload
-
-
-def _without_message_cache_control(raw_message: JsonValue, index: int) -> tuple[JsonValue, bool]:
-    """Drop a supported ``cache_control`` annotation from one Chat message.
-
-    Args:
-        raw_message: One ``messages`` entry.
-        index: Zero-based message index used in public error paths.
-
-    Returns:
-        The message (copied when an annotation is removed) and whether it changed.
-
-    Raises:
-        OpenAIProtocolError: The annotation is present but not a supported form.
-    """
-    if not isinstance(raw_message, dict):
-        return raw_message, False
-    message = cast(JsonObject, raw_message)
-    changed = False
-    if "cache_control" in message:
-        _require_supported_cache_control(
-            message["cache_control"], f"messages.{index}.cache_control"
-        )
-        message = {key: value for key, value in message.items() if key != "cache_control"}
-        changed = True
-    content = message.get("content")
-    if isinstance(content, list):
-        cleaned_content, content_changed = _without_text_part_cache_control(
-            cast(list[JsonValue], content), index
-        )
-        if content_changed:
-            if not changed:
-                message = dict(message)
-            message["content"] = cleaned_content
-            changed = True
-    return message, changed
-
-
-def _without_text_part_cache_control(
-    parts: list[JsonValue], message_index: int
-) -> tuple[list[JsonValue], bool]:
-    """Drop supported ``cache_control`` from OpenCode text content parts.
-
-    Args:
-        parts: Message ``content`` array.
-        message_index: Zero-based parent message index used in public error paths.
-
-    Returns:
-        The content array (copied when an annotation is removed) and whether it changed.
-
-    Raises:
-        OpenAIProtocolError: A text-part annotation is present but not a supported form.
-    """
-    cleaned: list[JsonValue] = []
-    changed = False
-    for part_index, raw_part in enumerate(parts):
-        if not isinstance(raw_part, dict) or "cache_control" not in raw_part:
-            cleaned.append(raw_part)
-            continue
-        part = cast(JsonObject, raw_part)
-        if part.get("type") not in _TEXT_PART_TYPES:
-            cleaned.append(raw_part)
-            continue
-        _require_supported_cache_control(
-            part["cache_control"],
-            f"messages.{message_index}.content.{part_index}.cache_control",
-        )
-        cleaned.append({key: value for key, value in part.items() if key != "cache_control"})
-        changed = True
-    return (cleaned, True) if changed else (parts, False)
-
-
-def _require_supported_cache_control(value: JsonValue, param: str) -> None:
-    """Accept null or a supported ephemeral ``cache_control`` object.
-
-    Args:
-        value: Raw ``cache_control`` annotation.
-        param: Public dotted field path used in the error.
-
-    Raises:
-        OpenAIProtocolError: The annotation is malformed or unsupported.
-    """
-    if value is None:
-        return
-    try:
-        _EphemeralCacheControl.model_validate(value)
-    except ValidationError as exc:
-        raise invalid_field(param) from exc
 
 
 def _validate_manifest(payload: JsonObject, manifest: CompatibilityManifest) -> None:
